feat: respect XDG user directories
Previously, on gnu/linux, the videos directory was determined by
combining $HOME with "VIDEOS", resulting in `$HOME/Videos`
However, since user directories can be customized using `xdg-user-dir`
(which is supported by almost all DE), hardcoding `$HOME/Videos` is not
ideal.

You can query user dirs by running `xdg-user-dir DIRNAME`. The one we're
interested in are `VIDEOS` and `MUSIC`. If the command fails (either the
user don't have it installed if they're using something like a WM, or
the user is not on gnu/linux) it will fallback to the previous
implementation.

For more information about `xdg-user-dir`, refer to the arch wiki:
https://wiki.archlinux.org/title/XDG_user_directories