feat: add `title` and date option for `filename_template` (#39)
* feat: add `title` option for `filename_template`

This option will return `media-title` of the file. If it's not present
however, it will fallback into `filename` and will apply the same
operation as `filename` (remove extension).

Why? Well if we're playing a video from http, `media-title` will be more
useful than `filename`. For example, when we play a youtube video, the
`filename` is `watch?v={video_id}` and the `media-title` would be the
actual title of the video.

* refactor

* feat: add date and time option under capital letters

This will add `%Y` as year, `%M` as month, `%D` as day, `%H` as hour
(24), `%I` as hour (12), `%P` as am/pm, `%N` as minute, `%S` as second

---------

Co-authored-by: Ren Tatsumoto <tatsu@autistici.org>